Rework GPIO routines to be clearer
It doesn't make sense to have a function to set the direction and one to
set the open drain status; it's clearer to have a single function that
sets the pin as an input and another that sets it as an output, with
the open drain status as a modifier on that call. That way when we're
changing modes we can be explicit about exactly what configuration we
program into the GPIO configuration register.