Add support for pull-up/power control
The stock Maple Mini doesn't support this, but work is in progress for a
daughterboard that would enable at least software-controlled pull-ups,
so add in the framework for this. Additionally wire it into the Linux
emulation code, outputting the state as part of the VCD and using it to
determine what result we return for a GPIO read.