Add short / pull-up detection for 1-Wire + I2C
These protocols rely on a pull-up resistor to pull the I/O line(s) high.
If it's missing or there's a short we'll potentially fail to read a 1
when we release the line. Add checks to detect this so we can try to
alert the user if this is the case.