Clean up main loop buffer / connection debug
The buffer in our main loop only needs to be CDC_BUFSIZE - the larger
size is from the initial sample code that printed out a hex string at
the start of every line. So trim it down, rename to a less generic name,
and change the debug connection string to actually indicate what we're
counting and not mention chopstx.